background: Fixed wrong ref
This commit fixes a comment from James on Rocketchat:

... i "Data set" afsnittet, skriver du: "Furthermore, specimens were
infiltrated..." Hvad er det, du citerer der? Det er vel ikke 40 år
siden, prøverne blev taget og præpareret?

Deleted the wrong references donath1982 and donath1993, which were
referring to the process (regarding MMA and such), and was actually
redundant and misleading. The correct reference to erben1997 has been
kept, but moved one sentence, because it is used as a reference in both
sentences.

Also wrapped text to textwidth=100 so it fits with the rest.