Merged changes from SN-main to individual files
The journal used for the last submission, required a single .tex file
with all sections. After we made this, we added further changes to the
text, figures, formatting, etc. This commit takes all these changes and
distributes them into their respective (old) individual files. This
gives a simpler workflow when working on many changes leading up to the
next submission.