Reorganized physical effects
Renamed to imaging artifacts instead, and provided clear definition of
how we use the words noise and artifacts and what they refer to.

Explained some parts more thoroughly and other parts less so, because it
is not relevant to go on about in depth. Keep it short as suggested by
reviewer.

Also changed old refs to sec:physics in metode.tex.