verification: changed figure 11
Now it uses subfigure with subcaptions to indicate what each
segmentation is, making it easier for readers to go back and forth and
compare the segmentations. Also a note about the streaking patterns seen
in the global threshold has been linked to the previous figure 2, which
shows the various physical effects.