published data: moved appendix to ERDA README
The appendix with information about published data, sizes, types of
volumes etc. was moved to a markdown file. It is currently only stored
in this repo, but the idea is that it becomes part of whatever ERDA
directory the data will become available at.

It was moved because it will only be relevant to understand this
information, if you also want to look at the data yourself. This is
independent of the code (which is more general than the data it is used
to process for the article).