adapted figure sizes
Most figures need larger ticks and labels so they can easily be read.
Figures with colorbars also need the size adapted to the figure size.
This work has been done for Figures 1-5+8+11+14+15. However the
remaining are yet to be fixed. This can however wait until after review.
There is a file for creating all article figures currently on TH2, and
this will be pushed to the evalmaxi dir.